Add operationType converter and AmountMapper

+++ src/main/java/com/challenge/transactions/dto/request/TransactionRequestDTO.java
@@ -14,7 +14,7 @@ import lombok.Setter;
 public class TransactionRequestDTO {
 
 	private Long accountId;
-	private Long operationTypeId;
+	private Integer operationTypeId;
 	private BigDecimal amount;
 
 }
